refactor: queue design
