+++ internal/kubernetes/manifest.go
@@ -107,8 +107,8 @@ func ApplyBuilderConsumerDeploymentManifest(dto ConfigureMaxWorkerDTO, config *c
 								accorev1.EnvVar().WithName("WORKER_IMAGE_URI").WithValue(config.Agent.WorkerImageUri),
 								accorev1.EnvVar().WithName("BACKEND_PORT").WithValue(strconv.FormatInt(int64(config.Backend.Port), 10)),
 								accorev1.EnvVar().WithName("BACKEND_AGENT_ENDPOINT").WithValue(config.Backend.AgentEndpoint),
-								accorev1.EnvVar().WithName("BUILDER_QUEUE_URI").WithValue(config.BuilderConfig.QueueUri),
-								accorev1.EnvVar().WithName("BUILDER_QUEUE_NAME").WithValue(config.BuilderConfig.QueueName),
+								accorev1.EnvVar().WithName("CONSUMER_QUEUE_URI").WithValue(config.ConsumerConfig.QueueUri),
+								accorev1.EnvVar().WithName("CONSUMER_ORGANIZATION_NAME").WithValue(config.ConsumerConfig.OrganizationName),
 							),
 					),
 				),
